requirements.txt
streamlit==1.32.0
google-genai==0.3.0
fpdf2==2.7.9
python-dotenv==1.0.1
pillow==10.2.0